Made changes to Entities and their respective controllers and services

+++ src/main/java/com/bhs/thinkbridge/services/UserService.java
@@ -7,6 +7,7 @@ import com.bhs.thinkbridge.repositories.UserRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,19 +29,17 @@ public class UserService {
 
     public Optional<?> createUser(UserDTO userDTO){
         if(userDTO.getUser_name() == null || userDTO.getEmail() == null){
-            ErrorDTO error = new ErrorDTO(HttpStatus.BAD_REQUEST);
-            error.setError("Both Username and Email are Required");
-            return Optional.of(error);
+            return Optional.of(new ErrorDTO(HttpStatus.BAD_REQUEST,"Both Username and Email are Required"));
         }
         if(userRepository.existsByEmailAndUserName(userDTO.getEmail(),userDTO.getUser_name())){
-            ErrorDTO error = new ErrorDTO(HttpStatus.BAD_REQUEST);
-            error.setError("Username or Email Already Exists!!! Please try with new one");
-            return Optional.of(error);
+            return Optional.of(new ErrorDTO(HttpStatus.BAD_REQUEST,"Username or Email Already Exists!!! Please try with new one"));
         }
         User newUser = User.builder()
                 .user_name(userDTO.getUser_name())
                 .email(userDTO.getEmail())
                 .bio(userDTO.getBio())
+                .created_at(new Date())
+                .updated_at(new Date())
                 .build();
         userRepository.save(newUser);
         return Optional.of(newUser);
@@ -48,12 +47,15 @@ public class UserService {
 
     public Optional<User> updateUser(String id,UserDTO userDTO){
         if(userRepository.existsById(id)){
-            userRepository.updateUser(id,userDTO.getUser_name(),userDTO.getEmail(),userDTO.getBio());
+            userRepository.updateUser(id,userDTO.getUser_name(),userDTO.getEmail(),userDTO.getBio(),new Date());
         }
         return userRepository.findById(id);
     }
 
     public void deleteUserById(String id){
-        userRepository.deleteById(id);
+        if(userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        }
     }
+    
 }
